Abilities moved to seperate file
Abilities enum ow in seperate file, PowerUp class now uses the enum value rather than string
